requirements.txt

Flask
Flask-SQLAlchemy
Flask-JWT-Extended
werkzeug
python-dotenv
Flask-Pydantic
pydantic
email-validator
gunicorn